+++ src/css/contact.css
@@ -30,7 +30,7 @@ b{
 }
 .timgc{
     max-width: 500px;
-    height: 400px;
+    height: 300px;
     margin-top: auto;
     margin-bottom: auto;
     padding-top: 20px;
